refactor(shapesMapCsg): * simplified parameters * moved reflect & groupValue functions to helpers * cleanups
